moved helpers to separate module, formatting
